Tidy some utility code
----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=111107107

+++ sdk/src/main/java/com/google/cloud/dataflow/sdk/util/VarInt.java
@@ -32,7 +32,7 @@ import java.io.OutputStream;
 public class VarInt {
 
   private static long convertIntToLongNoSignExtend(int v) {
-    return ((long) v) & 0xFFFFFFFFL;
+    return v & 0xFFFFFFFFL;
   }
 
   /**
